use * imports from backend and app packages
this is a little tricky, the release fails when these import from
0.0.0-semantically-released, but then yarn install hits npm registry
directly when they are marked as * and bypasses the local workspace.

not so concerned with the demo site right now so will leave as is. i
did a publish locally just removing these dependencies for a moment.

perhaps the issue is that the packages/* workspaces are being excluded
from the publish step